move private route to utils
